# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RLDB - API UNIVERSAL
  description: |
    API de Controle da Rede Logística Digital Brasileira (RLDB) - Integração Crossborder TR8
    
    Esta API funciona como um orquestrador neutro, permitindo que qualquer participante autorizado
    troque dados operacionais sem expor informações competitivas.
    
    A integração TR8 (Transporte Rodoviário Internacional de Cargas) permite operações crossborder
    entre Brasil e outros países, com conformidade total às regulamentações de comércio exterior.
  version: 1.0.0
  contact:
    name: RLDB API Support
    email: api-support@rldb.gov.br
    url: https://rldb.gov.br/api-support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.rldb.gov.br
    description: Production server
  - url: https://sandbox.rldb.gov.br
    description: Sandbox server for testing
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/crossborder/tr8/webhooks:
    post:
      tags:
        - TR8 Crossborder
      summary: Create webhook
      description: Creates a webhook configuration for TR8 events
      operationId: createWebhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrossborderWebhook'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrossborderWebhook'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - TR8 Crossborder
      summary: Get webhooks
      description: Retrieves webhook configurations for the current client
      operationId: getWebhooks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrossborderWebhook'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
...
    CrossborderWebhook:
      type: object
      properties:
        id:
          type: string
          description: Webhook ID
        client_id:
          type: string
          description: Client ID
        url:
          type: string
          description: Webhook URL
          example: "https://api.shopee.com/webhooks/rldb"
        events:
          type: array
          items:
            type: string
          description: Events to subscribe to
          example: ["tracking.updated", "shipment.delivered"]
        secret:
          type: string
          description: Webhook secret for HMAC verification
        active:
          type: boolean
          description: Whether webhook is active
          default: true
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
security:
  - bearerAuth: []